moved 3 functions to util.py
